--- src/components/Login/Login.jsx
@@ -1,1 +1,0 @@
-import * as React from 'react';
