+++ src/pages/Home/Home.css
@@ -1,0 +1,23 @@
+.search-window {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    transition:
+        background-color 0.3s,
+        color 0.3s;
+    padding: 0px 10px;
+}
+
+.light {
+    background-color: #f5f5f5;
+    color: #333;
+}
+
+.dark {
+    background-color: #333;
+    color: #f5f5f5;
+}
+
+.AppBar {
+    height: 60px;
+}
