+++ src/pages/Login/Login.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Login() {
+    return <div>Login</div>;
+}
+
+export default Login;
