+++ src/pages/Profile/Profile.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { useParams } from 'react-router';
+
+function Profile() {
+    const { uid } = useParams();
+    return <div>{uid}</div>;
+}
+
+export default Profile;
